src/modules/accounts/useCases/authenticateUser/AuthenticateUserUseCase.spec.ts
import { UnathorizedError } from "@errors/UnathorizedError";
import { ICreateUserDTO } from "@modules/accounts/dtos/ICreateUserDTO";
import { UsersRepositoryInMemory } from "@modules/accounts/repositories/in-memory/UsersRepositoryInMemory";

import { CreateUserUseCase } from "../createUser/CreateUserUseCase";
import { AuthenticateUseUseCase } from "./AuthenticateUserUseCase";

let usersRepositoryInMemory: UsersRepositoryInMemory;
let authenticateUserUseCase: AuthenticateUseUseCase;
let createUserUseCase: CreateUserUseCase;

describe("Authenticate user", () => {
  beforeEach(() => {
    usersRepositoryInMemory = new UsersRepositoryInMemory();
    authenticateUserUseCase = new AuthenticateUseUseCase(
      usersRepositoryInMemory
    );
    createUserUseCase = new CreateUserUseCase(usersRepositoryInMemory);
  });

  it("should be able authenticate a user", async () => {
    const user: ICreateUserDTO = {
      name: "carlos",
      email: "user@test.com",
      password: "gaby123",
      driver_license: "123456789",
    };
    await createUserUseCase.execute(user);
    const result = await authenticateUserUseCase.execute({
      email: user.email,
      password: user.password,
    });

    expect(result).toHaveProperty("token");
  });

  it("should not be able to authenticate a nonexistent user", () => {
    expect(async () => {
      await authenticateUserUseCase.execute({
        email: "aaaa",
        password: "aaaaa",
      });
    }).rejects.toBeInstanceOf(UnathorizedError);
  });

  it("should not be able authenticate a user with wrong password", async () => {
    const user: ICreateUserDTO = {
      name: "carlos",
      email: "user@test.com",
      password: "gaby123",
      driver_license: "123456789",
    };
    await createUserUseCase.execute(user);

    expect(async () => {
      await authenticateUserUseCase.execute({
        email: user.email,
        password: "incorrectPassword",
      });
    }).rejects.toBeInstanceOf(UnathorizedError);
  });
});
